+++ loaders/second.py
@@ -1,5 +1,6 @@
 from csv import DictReader
 from datetime import datetime
+from decimal import Decimal
 
 from unclebudget.models import Entry
 
@@ -8,7 +9,7 @@ def load(file):
     entries = []
     for row in DictReader(file):
         entries.append(Entry(
-            amount = float(row['Amount']),
+            amount = Decimal(row['Amount']),
             date = datetime.strptime(row['Transaction Date'], '%Y-%m-%d').date(),
             description = row['Transaction Detail'],
         ))
